roll back usage of REST api for bg generation
