+++ fresh.gen.ts
@@ -19,8 +19,7 @@ import * as $12 from './routes/home.tsx';
 import * as $13 from './routes/index.tsx';
 import * as $14 from './routes/resume.tsx';
 import * as $$0 from './islands/CalendlyWidget.tsx';
-import * as $$1 from './islands/GradientBackground.tsx';
-import * as $$2 from './islands/SearchBar.tsx';
+import * as $$1 from './islands/SearchBar.tsx';
 
 const manifest = {
 	routes: {
@@ -42,8 +41,7 @@ const manifest = {
 	},
 	islands: {
 		'./islands/CalendlyWidget.tsx': $$0,
-		'./islands/GradientBackground.tsx': $$1,
-		'./islands/SearchBar.tsx': $$2,
+		'./islands/SearchBar.tsx': $$1,
 	},
 	baseUrl: import.meta.url,
 	config,
